Move store code message builder to messages:: module
