+++ crates/cosmos/cosmos-relayer/src/contexts/chain.rs
@@ -1,7 +1,7 @@
 use alloc::sync::Arc;
 use core::ops::Deref;
 use hermes_wasm_test_components::components::WasmChainComponents;
-use hermes_wasm_test_components::traits::chain::store_code::StoreCodeMessageBuilderComponent;
+use hermes_wasm_test_components::traits::chain::messages::store_code::StoreCodeMessageBuilderComponent;
 use hermes_wasm_test_components::traits::chain::upload_client_code::{
     CanUploadWasmClientCode, WasmClientCodeUploaderComponent,
 };
