+++ crates/cosmos/cosmos-wasm-relayer/src/context/chain.rs
@@ -214,7 +214,7 @@ use hermes_runtime_components::traits::runtime::{
 };
 use hermes_test_components::chain::traits::queries::balance::CanQueryBalance;
 use hermes_wasm_test_components::components::WasmChainComponents;
-use hermes_wasm_test_components::traits::chain::store_code::StoreCodeMessageBuilderComponent;
+use hermes_wasm_test_components::traits::chain::messages::store_code::StoreCodeMessageBuilderComponent;
 use hermes_wasm_test_components::traits::chain::upload_client_code::{
     CanUploadWasmClientCode, WasmClientCodeUploaderComponent,
 };
