+++ crates/wasm/wasm-test-components/src/components.rs
@@ -1,8 +1,8 @@
 use cgp_core::prelude::*;
 
-use crate::impls::chain::store_code::BuildStoreCodeMessage;
+use crate::impls::chain::messages::store_code::BuildStoreCodeMessage;
 use crate::impls::chain::upload_client_code::SendStoreCodeProposalMessage;
-use crate::traits::chain::store_code::StoreCodeMessageBuilderComponent;
+use crate::traits::chain::messages::store_code::StoreCodeMessageBuilderComponent;
 use crate::traits::chain::upload_client_code::WasmClientCodeUploaderComponent;
 
 pub struct WasmChainComponents;
