+++ crates/wasm/wasm-test-components/src/impls/chain/messages/mod.rs
@@ -1,0 +1,1 @@
+pub mod store_code;
