+++ crates/wasm/wasm-test-components/src/traits/chain/mod.rs
@@ -1,2 +1,2 @@
-pub mod store_code;
+pub mod messages;
 pub mod upload_client_code;
